jwt done

+++ controllers/boaderController.js
@@ -1,17 +1,14 @@
 const asyncHandler = require("express-async-handler");
 const Boarders = require("../models/boarderModel");
+const User = require("../models/UserModel");
 /*
  @desc GET request
  @desc GET all the request
  @access PRIVATE
 */
 const getBoarders = asyncHandler(async (req, res) => {
-  try {
-    const boarders = await Boarders.find();
-    res.status(200).json(boarders);
-  } catch (error) {
-    throw new Error("Error getting the boarders", error);
-  }
+  const boarder = await Boarders.find({ user: req.user.id });
+  res.status(200).json(boarder);
 });
 
 /*
@@ -42,47 +39,36 @@ const getBoarder = asyncHandler(async (req, res) => {
 @access PRIVATE 
 */
 const addBoarder = asyncHandler(async (req, res) => {
-  try {
-    const { name, amount, room, due, starting } = req.body;
+  const { name, amount, room, due, starting } = req.body;
 
-    if (!name || !amount || !room || !due || !starting) {
-      return res.status(400).json({
-        error: "All fields are required",
-      });
-    }
+  if (!name || !amount || !room || !due || !starting) {
+    res.status(400);
+    throw new Error("All fields are required");
+  }
 
-    // Check if a boarder with the same room exists
-    const existingBoarder = await Boarders.findOne({ room });
+  // Check if a boarder with the same room exists
+  const existingBoarder = await Boarders.findOne({ room });
 
-    if (existingBoarder) {
-      if (
-        existingBoarder.name.toString() === name.toString() &&
-        existingBoarder.room.toString() === room.toString()
-      ) {
-        return res.status(400).json({
-          error: "Boarder already exists",
-        });
-      }
-    }
+  if (existingBoarder) {
+    res.status(400);
+    throw new Error("Boarder is already exist/room is occupied");
+  }
 
-    // Create a new boarder
-    const newBoarder = await Boarders.create({
-      name,
-      amount,
-      room,
-      due,
-      starting,
-    });
+  // Create a new boarder
+  const newBoarder = await Boarders.create({
+    user: req.user.id,
+    name,
+    amount,
+    room,
+    due,
+    starting,
+  });
 
-    if (newBoarder) {
-      return res.status(201).json({
-        boarder: newBoarder,
-      });
-    }
-  } catch (error) {
-    return res.status(500).json({
-      error: "Error adding a boarder",
-    });
+  if (newBoarder) {
+    res.status(201).json(newBoarder);
+  } else {
+    res.status(500);
+    throw new Error("Server Error");
   }
 });
 
@@ -93,39 +79,46 @@ const addBoarder = asyncHandler(async (req, res) => {
 */
 
 const updateBoarder = asyncHandler(async (req, res) => {
-  try {
-    const { name, amount, room, due, starting } = req.body;
-    const updatedData = {};
+  const { name, amount, room, due, starting } = req.body;
+  const updatedData = {};
 
-    if (name) updatedData.name = name;
-    if (amount) updatedData.amount = amount;
-    if (room) updatedData.room = room;
-    if (due) updatedData.due = due;
-    if (starting) updatedData.starting = starting;
+  if (name) updatedData.name = name;
+  if (amount) updatedData.amount = amount;
+  if (room) updatedData.room = room;
+  if (due) updatedData.due = due;
+  if (starting) updatedData.starting = starting;
 
-    const boarder = await Boarders.findById(req.params.id);
-    if (!boarder) {
-      return res.status(404).json({
-        error: "Boarder not found",
-      });
+  const boarder = await Boarders.findById(req.params.id);
+  if (!boarder) {
+    res.status(404);
+    throw new Error("Boarder is not exist");
+  }
+
+  const user = await User.findById(req.user.id);
+  //  check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found!");
+  }
+
+  if (boarder.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User is not Authorized");
+  }
+
+  const updatedBoarder = await Boarders.findByIdAndUpdate(
+    req.params.id,
+    updatedData,
+    {
+      new: true,
     }
+  );
 
-    const updatedBoarder = await Boarders.findByIdAndUpdate(
-      req.params.id,
-      updatedData,
-      {
-        new: true,
-      }
-    );
-
-    return res.status(201).json({
-      message: "Successfully updated the boarder",
-      boarder: updatedBoarder,
-    });
-  } catch (error) {
-    return res.status(500).json({
-      error: "Server error",
-    });
+  if (updatedBoarder) {
+    res.status(201).json(updatedBoarder);
+  } else {
+    res.status(500);
+    throw new Error("Server Error");
   }
 });
 /*
@@ -134,22 +127,28 @@ const updateBoarder = asyncHandler(async (req, res) => {
 @access PRIVATE
 */
 const deleteBoarder = asyncHandler(async (req, res) => {
-  try {
-    const boarder = await Boarders.findById(req.params.id);
-    if (boarder) {
-      const deleteBoarder = await Boarders.findByIdAndDelete(req.params.id);
-      if (deleteBoarder) {
-        return res.status(200).json({
-          message: "Boarder has been deleted",
-        });
-      }
+  const boarder = await Boarders.findById(req.params.id);
+  if (boarder) {
+    const user = await User.findById(req.user.id);
+    //  check for user
+    if (!user) {
+      res.status(401);
+      throw new Error("User not found!");
     }
-  } catch (error) {
-    res.status(500).json({
-      error: "Server error",
-    });
+
+    if (boarder.user.toString() !== user.id) {
+      res.status(401);
+      throw new Error("User is not Authorized");
+    }
+    const deleteBoarder = await Boarders.findByIdAndDelete(req.params.id);
+    if (deleteBoarder) {
+      return res.status(200).json({
+        message: "Boarder has been deleted",
+      });
+    }
   }
 });
+
 module.exports = {
   getBoarders,
   addBoarder,
